requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
fastmcp==2.0.0
python-dotenv==1.0.0
playwright==1.49.0
beautifulsoup4==4.12.2
requests==2.31.0 